refactor: class and test namespace has been updated
