+++ composer.json
@@ -5,7 +5,7 @@
     "license": "MIT",
     "autoload": {
         "psr-4": {
-            "LionDatabase\\": "src/LionDatabase/"
+            "Lion\\Database\\": "src/LionDatabase/"
         }
     },
     "autoload-dev": {
@@ -18,7 +18,7 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^10.4",
-        "lion/test": "^1.1",
-        "fakerphp/faker": "^1.23"
+        "fakerphp/faker": "^1.23",
+        "lion/test": "^1.4"
     }
 }
