+++ src/LionDatabase/Helpers/DriverTrait.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace LionDatabase\Helpers;
+namespace Lion\Database\Helpers;
 
 trait DriverTrait
 {
