+++ tests/DriverTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Tests;
 
-use LionDatabase\Driver;
-use LionDatabase\Drivers\MySQL;
-use LionTest\Test;
+use Lion\Database\Driver;
+use Lion\Database\Drivers\MySQL;
+use Lion\Test\Test;
 
 class DriverTest extends Test
 {
